split off set_eeg_reference as its own mixin & add SetChannelsMixin to Info
