+++ .git-blame-ignore-revs
@@ -1,1 +1,2 @@
 e81ec528a42ac687f3d961ed5cf8e25f236925b0  # black
+20737c0bd7c0dafee0a00caa1dba8ce573fd0f53  # move SetChannelsMixin between files
